infra: move all infra to modules and split control.
Moving all the infra building to modules allows for a more
controlled and selective approach when building resources. It
also allows us to split the control structure up, which means we
have smaller a blast radius when performing changes and each
tester can run their own stack without impacting others.

The jobs directory has been moved outside of Terraform, as this
will host all jobs and it not Terraform code.

The jrasell-b20b731 infra has been left as an example to copy and
use for other testing infra stacks.