infra: add Traefik ingress and InfluxDB monitoring jobs and infra resources (#10)
* infra: optionally let Nomad ALB expose Traefik and InfluxDB.

The Nomad ALB can now optionally be used to expose Traefik and
InfluxDB to the internet. This exposure is limited by the callers
public IP.

When left empty, the new variable is checked and the relevant
AWS resources are not created or modified.

* infra: expose Traefik and InfluxDB via ALB in eu-west-2 core.

This infra represents our core platform which will collect metrics
for bench infra. We therefore should expose Traefik and InfluxDB
to project members.

The output file has been updated to display commands for running
the jobs.

* infra: add Traefk and InfluxDB Nomad jobs.