infra: add InfluxDB Telegraf Ansible role. (#19)
This Ansible role lets us configure telegraf agent for metric
collection on our Nomad servers.

The new playbook is named such that we would expect each server
pool to use its own configuration. This is because they will send
to different buckets.

I have not found a good way to automatically detect the InfluxDB
API endpoint, so that will need to be modified on each new cluster
build.