infra: install and configure Nomad TLS via Ansible.
This change adds a new Ansible task to install and configure Nomad
TLS. It also allows for the configuration of a bash profile which
allows for the CLI to work from the host without additional setup.

The playbooks have been renamed slightly to differentiate the
clusters and will form the basis of additional work in this area.