ansible: adjust vars and set values from Terraform
Set Ansible host and group variables from values defined in Terraform to
keep them more consistent when values change and make the playbooks more
dynamic. Variables defined exclusively on Terraform are prefixed with
`terraform_` to make it clear what their source is.

The `nomad-cluster` module was expanded to accept additional Ansible
variables for the `server` and `client` groups, which will be helpful
when creating test clusters.

The `nomad_region` variable was being declared twice (in the `common`
and `nomad` roles) which can lead to unexpect results as Ansible doesn't
guarantee precedence in this case. Since the `nomad-tls` module doesn't
support specifying the region, and always uses `global` in the path,
using a variable to lookup the certificates could result in the
incorrect path.

Also fixes the format for the InfluxDB URL in the Telegraf
configuration and the load balancer IP list for clients.