core: use Terraform Cloud as backend
Since we're storing credentials for long-running infrastructure it may
be better to store them in a HashiCorp service instead of a random S3
bucket.

TFC also provides more visibility on changes and can be useful when
debugging problems with the rest of the team.

It also doesn't require AWS credentials from Doormat to be loaded into
the local machine.