core: deploy InfluxDB
Generate secrets for InfluxDB token and admin password and store them as
Nomad variables.

The token needs to be generated along with the `core` infrastructure
because it is used by Ansible to configure the Telegraf agents. This
token is then passed to the `core-nomad` configuration via an output.

The `core-nomad` starts the InfluxDB job using the values from the Nomad
variable.