core: add Terraform configuration for core Nomad (#59)
* core: add Terraform configuration for core Nomad

Add new root module under `infra/eu-west-2/core-nomad` to configure the
Nomad core cluster using Terraform.

The initial configuration starts the AWS EBS CSI volume plugin and
creates a volume that will be used by InfluxDB.

In order for Terraform Cloud to access the private Nomad server, the job
`shared/nomad/jobs/tfc-agent.nomad.hcl` runs a private TFC runner in the
workspace used to configure Nomad.

* core: deploy InfluxDB

Generate secrets for InfluxDB token and admin password and store them as
Nomad variables.

The token needs to be generated along with the `core` infrastructure
because it is used by Ansible to configure the Telegraf agents. This
token is then passed to the `core-nomad` configuration via an output.

The `core-nomad` starts the InfluxDB job using the values from the Nomad
variable.

* core: serve InfluxDB using TLS

Encrypt traffic to InfluxDB using the same TLS certificate used by
Nomad.

* influxdb: collect disk metrics and format config

In addition to disk IOPS it's also important to monitor disk usage to
make sure we don't run out of space in the core cluster.

Also format the Telegraf configuration file to make it more consistent.

* core: deploy Telegraf agent to Nomad clients

It's important that we also monitor metrics for clients in the core
cluster, specially the InfluxDB volume disk space.

* core: remove host volume configuration

We're using a CSI volume for InfluxDB so we don't need to create a host
volume anymore.

Update the `nomad` role to remove the host volume config file if none
are defined.