influx: ensure the service is always started
Ansible handlers have the following behaviours:
  - only run at the end of the playbook.
  - only run if a task that triggers it changes something (`changed`
    instead of `ok`).
  - only run on hosts that did not fail.

This means that, without an explicit task to enable and start the
service, InfluxDB may not run if the playbook fails after the config
file is created, even after running the playbook again.

This happens because the tasks that call the `restart_influxdb_telegraf`
handler are idempotent, meaning the will only run once. If a later task
fails, the host as whole is considered as failed for the run, and the
trigger is not called on them.

Running the playbook again has not effect since those tasks will not
change.

Forcing the service state signals to Ansible that it should always make
sure the service is running.