core: fix `NOMAD_CACERT` output value
The value of `path.module` is relative to where the file is defined, so
the output value will be `./tls/nomad-agent-ca.pem`, which does not work
when a `nomad` command is executed in any other directory.

With Terraform running in a remote runner, we can't use the `abspath`
function because it will resolved to a path inside the runner.

The `$PWD` environment variable will be expanded when running the
`export` commands.