nomad: allow setting arbitrary Nomad configuration
The `nomad_config_extra` variable receives a string that is written to a
high precedence configuration file, meaning it overwrite other
configuration.

This is useful when testing scenarios where agents need to be configured
in a particular way that is not exposed via Ansible variables.

For example, to disable scheduling works in a host it's possible to
write this to `host_vars/test-cluster-*.yaml`.

```
nomad_config_extra: |
  server {
    num_schedulers = 0
  }
```