nomad-load: Support Prometheus and JSON format output metrics. (#109)
This change modifies the HTTP server, so that it now supports
both JSON and Prometheus formatted metrics. This adds flexibility
and means we can scrape the endpoint and locally look at JSON data
rather than having to understand and parse the more confusing
Prometheus output.

The interval and retention values are currently hardcoded to a more
aggressive value. This seems OK initially, as we want to ingest
this data at a high frequency for accuracy. In the future, we may
want to make this configurable.