ansible: create collection to store reusable roles. (#135)
While the previous approach worked for a single repository, it did
not work for repo that might want to also use the Ansible roles.

Ansible solves the problem of a group of roles in a single repo
with collections. A collection allows importers to use the group
of Ansible resources in a similar manner to a single repo that
contains a role. This allows both this repo and others to use the
available Ansible resources, without duplication.

In order to make this change, the location of the Ansible roles
has moved in order to match the collection namespace and name. It
also removes the role path configuration and updates the playbooks
to use the namespace collection.